Rename commands.py to command_controller.py
This file will soon not have any commands
